Formatting, comments, and code structure
As I'm reviewing the code for the first time this commit was created to
address small formatting, comments, and code structure-related nits. Including the
following.

- formatting: Delete arbitrary breaks at the start of some funcs.

- config.ReadConifg: Simplify logic such that behavior does not change
and less file interactions are done. Ensure that when config file
changes a bad config file doesn't cause the control to exit. Instead it
should ignore the bad configuration.

- kubecluster.version: Rename to GetK8sVersion to make the function
purpose more clear.

- Support the ability to provide a local kubeconfig for development purposes and
fallback to incluster config when needed.