Allow namespaces to be assigned to remote endpoints
Previously, updates to all resources would be reported to all endpoints.
This change allows a list of namespaces to be assigned to a remote
endpoint such that only updates to resources in those namespaces will
be reported to that endpoint.  If a remote endpoint does not have any
namespaces assigned, updates for all resources from all namespaces will
be sent by default.

Signed-off-by: Richard Lander <lander2k2@protonmail.com>