Add resource data to delete event payload
The resource object data is currently not sent because, by the time the
processor pulls the event off the queue, there is no object to look up.
This change has the informer's handler put the resource's object data
on the queue with the rest of the event data (for delete events only) so
that it can be attached to the payload.

Signed-off-by: Richard Lander <landerr@vmware.com>