Add exec liveness probe capability
This change adds liveness checks at the processor and emitter.  Both
operations touch a file.  These files are, in turn, checked by a liveness
checking function which touches or removes a file "/healthy" which may
be used by a livenessProbe.  Example deployment manifest was updated to
include an example livenessProbe.

Signed-off-by: Richard Lander <landerr@vmware.com>