Add ability to ignore namespaces
This change adds a configuration option `ignoreNamespaces` which accepts
an array of namespace names that Quartermaster will ignore altogether.
If configured the informer will not put resources from the specified
namespaces on the queue for emission.

This change also removes the `namespaceSelector` from the InformerClient
as it wasn't being used and, if used, could lead to unpredictable
results with namespace ignoring.

Signed-off-by: Richard Lander <landerr@vmware.com>