Allow the use of security context for non-root
This change moves the cache and health check files to a /quartermaster
directory that can be mounted on an emptyDir and have a security context
that includes `fsGroup` and does not require root access to write files.

Also inluded and example of how to define a deployment to use an
`emptyDir` scratch volume for cache and health check files.

Signed-off-by: Richard Lander <landerr@vmware.com>